更新
现在执行preprocess不会报错了，前提是不中断程序。

+++ util/preprocess.py
@@ -22,15 +22,18 @@ def mk_dirs(path):
 
 
 def process(labels_path, images_path, data_path):
+    if not os.path.exists(images_path):
+        return
     with open(labels_path) as f:
         for line in f.readlines():
             img = images_path + '/' + line.split()[0]
             path = data_path + '/' + line.split()[1]
             shutil.move(img, path)
+    shutil.rmtree(images_path)
 
 
+# 创建文件夹 train 和 test
 mk_dirs(data_dir)
+# 划分训练集和测试集，根据数字将文件分到对应的文件夹中
 process(train_labels, train_images, data_dir + '/train')
 process(test_labels, test_images, data_dir + '/test')
-os.remove(train_images)
-os.remove(test_images)
